Extract custom popper to a new component
