+++ src/styles/components/_all.scss
@@ -13,3 +13,4 @@
 @import './widget';
 @import './tour';
 @import './issue-breadcrumbs.scss';
+@import './popper.scss';
