+++ src/styles/components/popper.scss
@@ -1,0 +1,11 @@
+.popper {
+  background-color: white;
+
+  color: $human-blue;
+  margin-bottom: 8px;
+  padding: 12px;
+  border-radius: 5px;
+  border: 1px solid $human-pink;
+
+  z-index: 103;
+}
