+++ src/styles/components/widget.scss
@@ -310,18 +310,6 @@ $trix-button--icon-size: 1.6em;
             }
           }
         }
-
-        .attach-popper {
-          background-color: white;
-
-          color: $human-blue;
-          margin-bottom: 8px;
-          padding: 12px;
-          border-radius: 5px;
-          border: 1px solid $human-pink;
-
-          z-index: 103;
-        }
       }
     }
   }
